everything but tests

+++ spec/models/food_truck_spec.rb
@@ -4,7 +4,7 @@ describe FoodTruck do
   describe "table" do
     it { should have_db_column(:name).of_type(:string) }
     it { should have_db_column(:description).of_type(:string) }
-    it { should have_db_column(:avg_rating).of_type(:integer) }
+    it { should have_db_column(:avg_rating).of_type(:decimal) }
     it { should have_db_column(:location).of_type(:string) }
     it { should have_db_column(:created_at).of_type(:datetime) }
     it { should have_db_column(:updated_at).of_type(:datetime) }
